requirements.txt
torchmetrics
torchnet
numpy
pandas
scipy
scikit-learn
matplotlib
seaborn
tqdm
timm
opencv-python
